Modify some existing files to fix the error
