+++ src/style.css
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/css2?family=Poppins&display=swap');
+
 * {
   margin: 0;
   padding: 0;
@@ -15,26 +17,27 @@
   display: flex;
   flex-direction: column;
   align-items: flex-start;
+  height: 80vh;
 }
 
-/* .score-box {
+.first-container {
   display: flex;
   flex-direction: column;
   align-items: flex-start;
-  border: 2px black solid;
-  padding: 5px;
-  width: 30rem;
-} */
+  height: 80vh;
+}
 
 h1 {
-  margin-top: 1em;
+  font-family: 'Poppins', sans-serif;
+  margin-top: 10px;
   font-size: 43px;
   text-align: center;
 }
 
 h3 {
-  margin-bottom: 30px;
-  font-size: 40px;
+  font-family: 'Poppins', sans-serif;
+  margin-bottom: 20px;
+  font-size: 35px;
 }
 
 input {
@@ -45,12 +48,8 @@ input {
 }
 
 h2 {
-  margin-bottom: 30px;
-  font-size: 40px;
-}
-
-.refresh {
-  margin-bottom: 30px;
+  font-family: 'Poppins', sans-serif;
+  font-size: 35px;
 }
 
 .name {
@@ -61,6 +60,7 @@ h2 {
   display: flex;
   justify-content: space-between;
   align-items: center;
+  width: 22em;
 }
 
 button {
@@ -69,4 +69,12 @@ button {
   border: 1px solid black;
   border-radius: 4px;
   cursor: pointer;
+}
+
+li {
+  list-style: none;
+  font-size: 20px;
+  margin-bottom: 5px;
+  border: 2px solid black;
+  padding: 5px 86px;
 }
